chore: move functions to looper
