+++ components/FormikFields/PaperInputPicker/Looper/index.js
@@ -1,9 +1,21 @@
+import _ from 'lodash';
+import * as React from 'react';
 import { useEffect } from 'react/cjs/react.development';
+import {
+  View
+} from 'react-native';
+import {
+  Button
+} from 'react-native-paper';
+
+import I18n from '../../../../modules/i18n';
 
 const Looper = ({
   data, config, numberedQestions,
-  setNumberedQuestions,
-  setQuestionsToRepeat,
+  setNumberedQuestions, questionsToRepeat,
+  setQuestionsToRepeat, additionalQuestions,
+  setAdditionalQuestions, translatedLabel,
+  loopsAdded, setLoopsAdded
 }) => {
   const {
     formikKey, fieldType, numberQuestionsToRepeat
@@ -19,6 +31,8 @@ const Looper = ({
       updateQuestionsToRepeat();
     }
   }, [numberedQestions]);
+
+  const [individualLoopsAdded, setIndividualLoopsAdded] = React.useState(0);
 
   // number all questions in config (for looping ccapabilities)
   const numberQuestionsInConfig = () => {
@@ -45,7 +59,43 @@ const Looper = ({
     });
   };
 
-  return null;
+  // add another loop of questions in correct position in form (additionalQuestions)
+  // update number of loops added (+1)
+  const addLoop = () => {
+    // updated formikKey to ensure uniqueness
+    let updatedQuestions = [];
+    questionsToRepeat.forEach((question) => {
+      const updatedQuestion = _.cloneDeep(question);
+      updatedQuestion.formikKey = `${updatedQuestion.formikKey}__loop${individualLoopsAdded}`;
+      updatedQuestions = updatedQuestions.concat(updatedQuestion);
+    });
+
+    setAdditionalQuestions(additionalQuestions.concat(updatedQuestions));
+    setLoopsAdded(loopsAdded + 1);
+    setIndividualLoopsAdded(individualLoopsAdded + 1);
+  };
+
+  const removeLoop = () => {
+    setLoopsAdded(loopsAdded - 1);
+    setIndividualLoopsAdded(individualLoopsAdded - 1);
+    setAdditionalQuestions(additionalQuestions.slice(0,
+      additionalQuestions.length - numberQuestionsToRepeat));
+  };
+
+  return (
+    <View>
+      <Button onPress={() => addLoop()}>
+        {I18n.t('paperButton.addAdditional')}
+        {translatedLabel}
+      </Button>
+      {individualLoopsAdded !== 0 && (
+      <Button onPress={() => removeLoop()}>
+        {I18n.t('paperButton.removePrevious')}
+        {translatedLabel}
+      </Button>
+      )}
+    </View>
+  );
 };
 
 export default Looper;
