+++ components/FormikFields/PaperInputPicker/index.js
@@ -1,4 +1,3 @@
-import _ from 'lodash';
 import { Spinner } from 'native-base';
 import * as React from 'react';
 import {
@@ -28,7 +27,7 @@ const PaperInputPicker = ({
   ...rest
 }) => {
   const {
-    label, formikKey, fieldType, sideLabel, numberQuestionsToRepeat
+    label, formikKey, fieldType, sideLabel
   } = data;
 
   const {
@@ -68,30 +67,6 @@ const PaperInputPicker = ({
   const [numberedQestions, setNumberedQuestions] = React.useState({});
   const [questionsToRepeat, setQuestionsToRepeat] = React.useState([]);
   const [additionalQuestions, setAdditionalQuestions] = React.useState([]);
-  const [individualLoopsAdded, setIndividualLoopsAdded] = React.useState(0);
-
-  // add another loop of questions in correct position in form (additionalQuestions)
-  // update number of loops added (+1)
-  const addLoop = () => {
-    // updated formikKey to ensure uniqueness
-    let updatedQuestions = [];
-    questionsToRepeat.forEach((question) => {
-      const updatedQuestion = _.cloneDeep(question);
-      updatedQuestion.formikKey = `${updatedQuestion.formikKey}__loop${individualLoopsAdded}`;
-      updatedQuestions = updatedQuestions.concat(updatedQuestion);
-    });
-
-    setAdditionalQuestions(additionalQuestions.concat(updatedQuestions));
-    setLoopsAdded(loopsAdded + 1);
-    setIndividualLoopsAdded(individualLoopsAdded + 1);
-  };
-
-  const removeLoop = () => {
-    setLoopsAdded(loopsAdded - 1);
-    setIndividualLoopsAdded(individualLoopsAdded - 1);
-    setAdditionalQuestions(additionalQuestions.slice(0,
-      additionalQuestions.length - numberQuestionsToRepeat));
-  };
 
   return (
     <>
@@ -534,21 +509,8 @@ const PaperInputPicker = ({
         </View>
       )}
       {fieldType === 'loop' && (
-        <View>
-          <Looper
-            data={data}
-            config={config}
-            numberedQestions={numberedQestions}
-            setNumberedQuestions={setNumberedQuestions}
-            questionsToRepeat={questionsToRepeat}
-            setQuestionsToRepeat={setQuestionsToRepeat}
-            additionalQuestions={additionalQuestions}
-            setAdditionalQuestions={setAdditionalQuestions}
-            individualLoopsAdded={individualLoopsAdded}
-            setIndividualLoopsAdded={setIndividualLoopsAdded}
-          />
-          <View key={formikKey}>
-            {additionalQuestions !== undefined && additionalQuestions.length !== 0
+      <View key={formikKey}>
+        {additionalQuestions !== undefined && additionalQuestions.length !== 0
               && additionalQuestions.map((question) => (
                 <PaperInputPicker
                   data={question}
@@ -562,18 +524,20 @@ const PaperInputPicker = ({
                   setScrollViewScroll={setScrollViewScroll}
                 />
               ))}
-            <Button onPress={() => addLoop()}>
-              {I18n.t('paperButton.addAdditional')}
-              {translatedLabel}
-            </Button>
-            {individualLoopsAdded !== 0 && (
-            <Button onPress={() => removeLoop()}>
-              {I18n.t('paperButton.removePrevious')}
-              {translatedLabel}
-            </Button>
-            )}
-          </View>
-        </View>
+        <Looper
+          data={data}
+          config={config}
+          numberedQestions={numberedQestions}
+          setNumberedQuestions={setNumberedQuestions}
+          questionsToRepeat={questionsToRepeat}
+          setQuestionsToRepeat={setQuestionsToRepeat}
+          additionalQuestions={additionalQuestions}
+          setAdditionalQuestions={setAdditionalQuestions}
+          translatedLabel={translatedLabel}
+          loopsAdded={loopsAdded}
+          setLoopsAdded={setLoopsAdded}
+        />
+      </View>
       )}
     </>
   );
